# openapi/task_openapi.yaml
openapi: 3.0.0
info:
  title: Task Management API
  version: 1.0.0
paths:
  /tasks/{user_id}:
    get:
      summary: Get tasks for a specific user
      tags:
        - tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: integer
      responses:
        '200':
          description: A list of tasks for the specified user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '404':
          description: User not found or no tasks for this user
  /tasks:
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTask'
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    patch:
      summary: Update an existing task
      tags:
        - tasks
      requestBody:
        description: The task with updated details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTask'
      responses:
        '200':
          description: The updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      tags:
        - tasks
      requestBody:
        description: The task to delete
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteTask'
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer

    CreateTask:
      type: object
      properties:
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
      required:
        - task
        - user_id

    UpdateTask:
      type: object
      properties:
        id:
          type: integer
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
      required:
        - id
        - task

    DeleteTask:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
      required:
        - id
        - user_id